requirements.txt
python-telegram-bot==21.10
matplotlib==3.9.0
Pillow==10.4.0
pytest==8.3.2
pytest-asyncio==0.24.0
